Initial compiling migration
