+++ core/lyo-core-model/pom.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>org.eclipse.lyo.oslc4j.core</groupId>
+    <artifactId>oslc4j-core-build</artifactId>
+    <version>5.0.1-SNAPSHOT</version>
+    <relativePath>../oslc4j-core-build/pom.xml</relativePath>
+  </parent>
+  <name>Lyo :: Core :: OSLC Core model</name>
+  <description>Core model of the Eclipse Lyo OSLC4J SDK.</description>
+
+  <artifactId>lyo-core-model</artifactId>
+
+  <properties>
+  </properties>
+
+</project>
