+++ core/lyo-core-model/src/main/java/org/eclipse/lyo/oslc4j/core/model/Occurs.java
