+++ pom.xml
@@ -86,8 +86,8 @@
     <module>trs/server</module>
     <module>trs/client</module>
     <module>store</module>
-    <!-- Remove util https://bugs.eclipse.org/bugs/show_bug.cgi?id=574043
-    and/or https://github.com/weso/shaclex/issues/390 are resolved-->
+      <!-- Remove util https://bugs.eclipse.org/bugs/show_bug.cgi?id=574043
+      and/or https://github.com/weso/shaclex/issues/390 are resolved-->
     <!--    <module>validation</module>-->
   </modules>
